fix docstring and rename file
