+++ configs/sot/siamese_rpn/siamese_rpn_r50_1x_uav123.py
@@ -9,9 +9,9 @@ data_root = 'data/'
 data = dict(
     val=dict(
         type='UAV123Dataset',
-        ann_file=data_root + 'uav123/annotations/uav123.json',
-        img_prefix=data_root + 'uav123/data_seq/UAV123'),
+        ann_file=data_root + 'UAV123/annotations/uav123.json',
+        img_prefix=data_root + 'UAV123/data_seq/UAV123'),
     test=dict(
         type='UAV123Dataset',
-        ann_file=data_root + 'uav123/annotations/uav123.json',
-        img_prefix=data_root + 'uav123/data_seq/UAV123'))
+        ann_file=data_root + 'UAV123/annotations/uav123.json',
+        img_prefix=data_root + 'UAV123/data_seq/UAV123'))
