+++ docs/dataset.md
@@ -94,7 +94,7 @@ python ./tools/convert_datasets/mot2coco.py -i ./data/MOT17/ -o ./data/MOT17/ann
 python ./tools/convert_datasets/mot2reid.py -i ./data/MOT17/ -o ./data/MOT17/reid --val-split 0.2 --vis-threshold 0.3
 
 # UAV123
-python ./tools/convert_datasets/uav2coco.py -i ./data/uav123/ -o ./data/uav123/annotations
+python ./tools/convert_datasets/uav2coco.py -i ./data/UAV123/ -o ./data/UAV123/annotations
 ```
 
 The folder structure will be as following after your run these scripts:
