+++ docs/useful_tools_scripts.md
@@ -111,15 +111,16 @@ python tools/publish_model.py work_dirs/dff_faster_rcnn_r101_dc5_1x_imagenetvid/
 
 The final output filename will be `dff_faster_rcnn_r101_dc5_1x_imagenetvid_20201230-{hash id}.pth`.
 
-### SiameseRPN++ Test-time Parameter Search
+## SiameseRPN++ Test-time Parameter Search
 
-`tools/sot_siamrpn_param_search.py` can search the test-time tracking parameters in SiameseRPN: `penalty_k`, `lr` and `window_influence`. You need to pass the searching range of each parameter into the argparser.
+`tools/sot_siamrpn_param_search.py` can search the test-time tracking parameters in SiameseRPN++: `penalty_k`, `lr` and `window_influence`. You need to pass the searching range of each parameter into the argparser.
 
 Example on UAV123 dataset:
 
 ```shell
-python tools/sot_siamrpn_param_search.py [${CONFIG}] [--checkpoint ${CHECKPOINT}] [--penalty-k-range 0.05,0.5,0.05]
-[--lr-range 0.3,0.45,0.02] [--win-infu-range 0.46,0.55,0.02] [--log ${LOG}] [--eval ${EVAL}]
+./tools/dist_sot_siamrpn_param_search.sh [${CONFIG_FILE}] [$GPUS] \
+[--checkpoint ${CHECKPOINT}] [--log ${LOG_FILENAME}] [--eval ${EVAL}] \
+[--penalty-k-range 0.05,0.5,0.05] [--lr-range 0.3,0.45,0.02] [--win-infu-range 0.46,0.55,0.02]
 ```
 
 ## Miscellaneous
