+++ tools/convert_datasets/uav2coco.py
@@ -32,9 +32,9 @@ def convert_uav123(uav123, ann_dir, save_dir):
         ann_dir (str): The path of trackingnet test dataset
         save_dir (str): The path to save `uav123`.
     """
-    # The format of each line in "uav_info.txt" is
+    # The format of each line in "uav_info123.txt" is
     # "anno_name,anno_path,video_path,start_frame,end_frame"
-    info_path = osp.join(os.path.dirname(__file__), 'uav_info.txt')
+    info_path = osp.join(os.path.dirname(__file__), 'uav123_info.txt')
     uav_info = mmcv.list_from_file(info_path)[1:]
 
     records = dict(vid_id=1, img_id=1, ann_id=1, global_instance_id=1)
