+++ tools/dist_sot_siamrpn_search.sh
@@ -1,0 +1,8 @@
+CONFIG=$1
+GPUS=$2
+PORT=${PORT:-29500}
+
+
+PYTHONPATH="$(dirname $0)/..":$PYTHONPATH \
+python -m torch.distributed.launch --nproc_per_node=$GPUS --master_port=$PORT \
+    $(dirname "$0")/sot_siamrpn_param_search.py $CONFIG --launcher pytorch ${@:3}
